Removed Protobufs
libprotobuf being 32 bit only caused problems with this.
All protobufs are in a seperate repository that users can clone and
build and handle upon their needs. It is recomended that they have a
standalone 64 bit build of it and communicate that somehow to the 32 bit
steam library to prevent contamination between custom build and system
installed libraries, and to prevent a whole sphagheti of mess when
building.